Code simplification
DRY fixes in views.py - a lot of code moved to services.py
