+++ app/urls.py
@@ -1,5 +1,5 @@
-from django.urls import path
 from .views import *
+from django.urls import path
 
 urlpatterns = [
     path('', index_page, name='index_page_url'),
