+++ app/views.py
@@ -1,13 +1,7 @@
-from django.shortcuts import render, redirect, get_object_or_404
-from .forms import *
+from .services import *
+from django.shortcuts import render, redirect
 from django.http import JsonResponse
 from .ascii_generators import ascii_generators
-from PIL import Image
-from django.conf import settings
-import os
-from django.urls import reverse
-from django.core.files import File
-from django.utils.translation import gettext_lazy as _
 
 
 def handler400_view(request, *args, **kwargs):
@@ -29,8 +23,9 @@ def index_page(request, app_txt_mode=False, ascii_obj=None):
         'app_txt_mode': app_txt_mode,  # indicates if user is opening "Text to ASCII"
     }
 
+    # if ascii_obj is passed in, add report form to context
     if ascii_obj is not None:
-        context['report_form'] = ReportForm()
+        context['report_form'] = ReportService.get_form()
 
     return render(request, 'app/index.html', context=context)
 
@@ -59,25 +54,22 @@ def sitemap_xml(request):
 def feedback(request):
     if request.is_ajax():
         if request.method == 'POST':
-            form = FeedbackForm(data=request.POST)
-            if form.is_valid():
-                form.save()
-                return JsonResponse({}, status=200)
-            return JsonResponse({'errors': form.errors}, status=400)
+            form, status = FeedbackService.create(request.POST)
+            response = {}
+            if status == 400:
+                response['errors'] = form.errors
+            return JsonResponse(response, status=status)
         return JsonResponse({}, status=405)  # 405 Method Not Allowed
     else:
         if request.method == 'GET':
-            form = FeedbackForm()
-            context = {'form': form}
-            return render(request, 'app/feedback.html', context=context)
+            form = FeedbackService.get_form()
+            return render(request, 'app/feedback.html', context={'form': form})
 
-        if request.method == 'POST':
-            form = FeedbackForm(request.POST)
-            context = {'form': form}
-            if form.is_valid():
-                form.save()
+        elif request.method == 'POST':
+            form, status = FeedbackService.create(request.POST)
+            if status == 200:
                 return redirect('feedback_url')
-            return render(request, 'app/feedback.html', context=context)
+            return render(request, 'app/feedback.html', context={'form': form})
 
 
 def image_to_ascii_generator(request):
@@ -106,97 +98,16 @@ def text_to_ascii_generator(request):
 
 
 def ascii_detail(request, ascii_url_code):
-    ascii_obj = get_object_or_404(GeneratedASCII, url_code=ascii_url_code)
-    try:
-        _ = ascii_obj.image_to_ascii_type
-        app_txt_mode = False
-    except ImageToASCIIType.DoesNotExist:
-        app_txt_mode = True
+    ascii_obj = GeneratedASCIIService.get_object_or_404(ascii_url_code)
+    app_txt_mode = GeneratedASCIIService.is_txt_mode(ascii_obj)
     return index_page(request, ascii_obj=ascii_obj, app_txt_mode=app_txt_mode)
-
-
-def _generate_unique_image_path(file_extension, r=0, r_max=10):
-    """
-    Recursive function that generates random unique file name and path.
-    :return: Full path to file, name with extension.
-    """
-    name = ''.join(random.choices(string.ascii_lowercase + string.digits, k=50))
-    y = f'{name}{file_extension}'
-    x = os.path.join(settings.MEDIA_ROOT, 'input_images/', y)
-    if os.path.exists(x):
-        if r >= r_max:
-            return None, None
-        x, y = _generate_unique_image_path(file_extension, r=r + 1)
-    return x, y
 
 
 def ascii_share(request):
     if request.is_ajax():
         if request.method == 'POST':
-            if request.POST.get('file_name', False):
-                img2ascii_mode = True
-                result = ascii_generators.image_to_ascii_generator(request)
-            else:
-                img2ascii_mode = False
-                result = {
-                    'txt': request.POST.get('txt', ''),
-                    'txt_multi': request.POST.get('txt_multi', ''),
-                    'arts': ascii_generators.text_to_ascii_generator(request)
-                }
-            if type(result) == JsonResponse:
-                return result
-            preferred_output_method = request.POST.get('preferred_output_method', None)
-            if preferred_output_method is None:
-                return JsonResponse({}, status=400)
-            ascii_obj = GeneratedASCII.objects.create(preferred_output_method=preferred_output_method)
-            if img2ascii_mode:
-                image_to_ascii_type_obj = ImageToASCIIType(
-                    generated_ascii=ascii_obj,
-                )
-                file = open(os.path.join(settings.TEMPORARY_IMAGES, result.get('file_name')), 'rb')
-                unused_fn, file_extension = os.path.splitext(file.name)
-                unused_path, file_name = _generate_unique_image_path(file_extension)
-                image_to_ascii_type_obj.input_image.save(
-                    file_name,
-                    File(file),
-                )
-                image_to_ascii_type_obj.save()
-                file.close()
-                ImageToASCIIOptions.objects.create(
-                    image_to_ascii_type=image_to_ascii_type_obj,
-                    columns=result.get('num_cols'),
-                    brightness=result.get('brightness'),
-                    contrast=result.get('contrast'),
-                )
-                arts = result.get('arts')
-                for i in range(len(arts)):
-                    OutputASCII.objects.create(
-                        generated_ascii=ascii_obj,
-                        method_name=str(i + 1),
-                        ascii_txt=arts[i]
-                    )
-            else:
-                if not request.POST.get('multiple_strings', False):
-                    multi_line_mode = False
-                    input_text = result.get('txt', '')
-                else:
-                    multi_line_mode = True
-                    input_text = result.get('txt_multi', '')
-                TextToASCIIType.objects.create(
-                    generated_ascii=ascii_obj,
-                    input_text=input_text,
-                    multi_line_mode=multi_line_mode,
-                )
-                arts = result.get('arts')
-                for art in arts:
-                    OutputASCII.objects.create(
-                        generated_ascii=ascii_obj,
-                        method_name=art[0],
-                        ascii_txt=art[1],
-                    )
-            return JsonResponse({
-                'shared_redirect_url': reverse('ascii_detail_url', kwargs={'ascii_url_code': ascii_obj.url_code})
-            }, status=200)
+            json_response = GeneratedASCIIService.create(request)
+            return json_response
         return JsonResponse({}, status=405)
     else:
         return redirect('index_page_url')
@@ -205,17 +116,8 @@ def ascii_share(request):
 def ascii_report(request, ascii_url_code):
     if request.is_ajax():
         if request.method == 'POST':
-            form = ReportForm(data=request.POST)
-            try:
-                generated_ascii = GeneratedASCII.objects.get(url_code=ascii_url_code)
-            except GeneratedASCII.DoesNotExist:
-                return JsonResponse({'errors': {'captcha': [_("Can't find object to report.")]}}, status=400)
-            if form.is_valid():
-                Report.objects.create(text=form.cleaned_data['text'],
-                                      email=form.cleaned_data['email'],
-                                      generated_ascii=generated_ascii)
-                return JsonResponse({}, status=200)
-            return JsonResponse({'errors': form.errors}, status=400)
+            json_response = ReportService.create(data=request.POST, ascii_url_code=ascii_url_code)
+            return json_response
         return JsonResponse({}, status=405)  # 405 Method Not Allowed
     else:
         return redirect('index_page_url')
